Query for GPU fence support in passthrough CMD decoder
Need to query for the capability otherwise the browser process won't
create a fence object and pass it in for presentation. The GPU process
expects this fence as the synchronization primitive when supported. If
it isn't provided the submitted buffer may be presented before the GPU
finishes rendering into it.

BUG=1097320,1083171
TEST=Manually on device and checked for tearing with passthrough command
decoder.

Change-Id: Ic1449d0fa0c0178078a88ac78fb38d94d950a11b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2300810
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Daniel Nicoara <dnicoara@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789086}