media: Remove expiration update in AesDecryptor
The EME spec says that the expiration time should stay NaN if no such
time exists or if the license explicitly never expires, which is the
case for Clear Key. Therefore, AesDecryptor never needs to update the
expiration. Currently we didn't update it in UpdateSession(), and the
update in RemoveSession() pretty much has no effect. To avoid confusion,
this CL removes the call in RemoveSession() as well.

Change-Id: I3439ef9ecb31b5b788880a2b352f45b0bf494030
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2297692
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789146}