media: Replace PlayerTracker with CdmContext::EventCB
PlayerTracker used to serve two purposes: notify when new usable key is
available and when the CDM is destructed. The latter is not needed since
the CDM will always outlive the media element now (in fact we pass
base::DoNothing() for that). The standard way to notify new key is via
CdmContext::EventCB.

This CL removes PlayerTracker and use CdmContext::EventCB for new key
notification, as the rest of media pipeline does.

Bug: 413413,821288
Change-Id: I08b81a8e63902632d516bb8542b5b70d42038afb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2296268
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789152}