[omnibox] Unbold all ZeroSuggest search results
We currently classify ZeroSuggest search results against the current
omnibox text. This was a holdover from the Query in Omnibox experiment,
when we only did ZeroSuggest against the displayed query on SERP.

Now that Query in Omnibox is completely dead, and we are doing
ZeroSuggest against arbitrary page URLs, we want to change the behavior
to never classify ZeroSuggest search results, and leave them all
unbolded.

This patch is minimal and designed to be mergeable back to M85.

Bug: 1095837, 1106096
Change-Id: Icceec38d9d8697b69999115af4db9ae3dabd32d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2300990
Commit-Queue: Tommy Li <tommycli@chromium.org>
Commit-Queue: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789162}