webview: alternate option for extra headers.
Implement an alternative option for the handling of extra headers in
loadUrl, which only allows them to be sent to redirects to the same
domain (eTLD+1). This is less restrictive than the existing feature flag
requiring that they be the same origin (and if that flag is enabled,
this one has no effect).

Implement a new histogram that just records whether redirects are
same-origin, same-domain, or cross-domain, and obsolete the current one,
which doesn't make sense with the new option.

Bug: 1038002
Change-Id: I121de069d7a87fb4718ad9131ca427a5c017a7d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264609
Auto-Submit: Richard Coles <torne@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789174}