webauthn: maybe fix Fido2CredentialRequestTest
Fido2CredentialRequestTest in chrome_public_test_apk is failing. Commit
f55cd37e32 was reverted[1] but I suspect that the cause might actually
be the Mojo change in c296f39e2a. Trying to run the tests fails for me
because the tests timeout trying to do setup as I'm remote, but I think
this is worth a shot at greening the bots.

TBR=martinkr

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2303030

Change-Id: Ida3e66df0c5bd9e98ffc2b84032d857236aaf6e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303354
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#789184}