Move ChromeInstallerCleanup checking into mini_installer::Configuration.
The debugging bit that controls whether or not the extracted files are
left behind rather than cleaned up is a kind of configuration, so it's
natural that it reside in the Configuration class. This is a refactor
with no behavior change with one exception: this CL fixes a regression
introduced in r282032.

BUG=None

Change-Id: I6dd594d220b9b75a507f4482af49f083b697b3d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302152
Commit-Queue: Greg Thompson <grt@chromium.org>
Reviewed-by: Sébastien Marchand <sebmarchand@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789193}