Cros: Update paging after a drag reparenting is cancelled
While repareting an app from a folder into the main grid, we could drag
it to create a new blank page. If the repareting is cancelled, the app
will snap back to the folder but the blank page will remain, causing the
system to crash eventually.
We need to update the apps grid paging in this case, so that the blank
pages are discarded.

Bug: 1100011
Change-Id: I5e28ed61bb50f3bde85cceaca9c5c7d0be1693b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2300836
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789202}