Uncouple PDF printing metafile usage from plugin
Printing from within the context of a pepper plugin for a Flash object
started crashing once the print composite usage stopped the inefficient
practice of generating metafiles for both individual pages and the full
document, in r751614.

Even though this issue of printing a Flash object is not of much
interest because of Flash being near end-of-life, we can at least not
crash the tab on users who happen to invoke such a print.

This restores ability of print preview to complete and provide page
images, although the content of the Flash object is empty.  This at
least matches the behavior prior to r751614.

Bug: 1098860
Change-Id: I86f7030ade0f126d6994b810f26775bf8cfe2693
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298036
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789239}