[Switch Access] Listen to locationChanged on all ancestors
When a window is resized, a locationChanged event is not fired on its
descendants (e.g. the maximize button, the web contents). So Switch
Access' locationChanged handler should listen to events on all of the
current node's ancestors as well.

Also adds support for listening to multiple nodes with
RepeatedEventListener.

AX-Relnotes: n/a.
Bug: 1105296
Change-Id: Idba74866ef71fa0adebb1e08647a860404afb43e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2300825
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#789248}