content: fix GestureListenerManagerTest
This test is still mildly flaky. The issue is that while it waits
for a non-empty paint, this is not always enough. This makes the
test additionally wait for loading to complete and more importantly,
waits for a visual update. The latter is a good signal the renderer
will actually process events.

BUG=1091129
TEST=test only change

Change-Id: Iabc6d2dce83985086c56db916c8cc273e3ed65f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303554
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789280}