Add metrics for tracking scanning backends
Adds the DocumentScan.SaneBackend enum, as well as the
DocumentScan.{ScanRequested,ScanSucceeded,ScanFailed} metrics. These are
used to track which SANE backend a user requested when a scan is
started, as well as the backend used when a scan finished successfully
or unsuccessfully.

Metrics are added by https://crrev.com/c/2281282

Obsoletes the ScanResult metrics since it is superseded by the new
metrics.

Bug: b:160472550
Change-Id: I5eae8fb99ee7d09ce39688648a6b6828d9122190
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2281207
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789297}