Ad metrics: Add |earliest_first_contentful_paint_| to fix FEtP crash.
Currently, the AdsPageLoadMetricObserver only records the FCP of the
root ad frame in FrameData and ignores FCP of any child frames except
for an ephemeral check while computing the Creative Origin Status.

This can lead to an situation where a child frame receives an
First Eligible to Paint and then an FCP, allowing calculation of
the Creative Origin Status, but the root ad frame has not yet
received an FCP, so the FrameData's timing object still
holds a null value for the FCP. Then, if a request to nullify the
First Eligible to Paint value for the frame tree arrives, under the
current setup it will succeed, and then a DCHECK for a non-null FEtP
will subsequently fail.

Further changes to fix the related bug in the FCP ads metric will be
in a follow-up CL.

Bug: 1103610,1103782
Change-Id: I4234ada8a6d0d98e2cf153522bd15e0d86efd8dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290775
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789300}