DCHECK fix for Legacy TableCell
Does not fix the layout, but it prevents the DCHECK.
Prcentage padding can cause negative widths in
Legacy table layout.
With TablesNG shipping soon, it is not worth
investigating root cause. This clamp just
prevents the DCHECK, and is not going to make
anything look worse.

Bug: 1105900
Change-Id: Id68369dbb906fa62238c71f83b7b87ac3743917a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303564
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Aleks Totic <atotic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789355}