Disabling testResumeInterruptedDownload.
This is a noop since UseDownloadOfflineContentProvider is set as
enabled in fieldtrial_testing_config.json, so
useDownloadOfflineContentProvider() always returns true and exits
the test immediately.

Disabling the test to prevent test failure on release testers, where
fieldtrial_testing_config.json is not used and causes
this test to fail.

If this test is needed, can it be fixed up and re-enabled.

Bug: 1106460
Change-Id: I0986131275f490ec9e75b26790e9aaaa07f172ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303787
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Ben Mason <benmason@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789357}