[IntentHandling] Upgrade stored preferred apps when sharing flag is on.
To support sharing, we added action condition in the intent filter.
However this will be inconsistent with the stored preferred apps
(contains intent filters). This will cause filter matching issues when
we try to modify or delete preferred apps entry. This CL adds the
support to upgrade the preferred apps if the flag is on.

BUG=1092784

AppServiceImplTest.PreferredAppsUpgrade.

Test: Unit test PreferredAppsConverterTest.UpgradePreferredApp and
Change-Id: Ib6fb3e4d6128b7c2f83a63cac73d800993a4b440
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2301549
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789360}