[ozone/wayland] Introduce output queue to WaylandBufferManagerHost.
The current implementation only allows two buffers in flight at a time.
This is because a new buffer may only be submitted when an old one is
released, and initially clients of WBMH only submit the initial buffer
plus one more (because of the immediate ack of the initial buffer).

Further, only submitting a new buffer when an old buffer is released
means that latency variation in receiving release messages directly
translates into increased latency in submitting the next buffer. Under
load when running linux Chrome on ChromeOS, IPC shunting release
messages back from viz=>exo=>ozone can have latencies in the low tens of
milliseconds. If buffers are shown in a separate DRM plane (i.e. not
copied) they must be held onto for the duration of scanout. This plus
the IPC release latency causes visible performance degredation.
Currently this latency cannot be absorbed by the buffer queue in
linux Chrome.

This change introduces an output queue in WBMH. Note that this does not
change the semantics of OnSubmission and OnPresentation (acking). For
this change to be useful, clients (e.g. GbmSurfacelessWayland) also need
to have an output queue, which I will add in a follow-up CL.

In order to provide the guarantees swap ack expects this CL also adds
some logic for destroyed buffers and out of order releases. For linux
Chrome on ChromeOS, surface sync logic in Viz (in ChromeOS chrome) can
cause buffer releases to be re-ordered - the new logic handles this case.

Bug: 1062529
Test: added unit tests
Test: Run linux chrome on chromeos
Test: Run ozone linux chrome on Weston, software and hw accel
Change-Id: I3574248c92a8cf12207154052194c0e021d2984f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2287334
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789375}