Transferable Streams: Don't permit the remote side to crash us
The transferable streams implementation assumed that two kPull messages
couldn't be sent without an intervening chunk. This wasn't true in the
case of a compromised renderer, and since
https://chromium-review.googlesource.com/c/chromium/src/+/2291180 hasn't
been true in general.

This change aligns the implementation with the proposed standard at
https://chromium-review.googlesource.com/c/chromium/src/+/2291180.

This fixes flakiness in the http/tests/streams/transferable/worker.html
test.

BUG=1106239,1106239

Change-Id: I12c9d011c4a42c8178fcaffa4ddd23e38fb82d9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303329
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789389}