[Autofill Assistant] Do not revive when switching from CCT to Browser
Before this change, when the Autofill Assistant was dismissed with the
cancel button showing the Snackbar, and changing from CCT to Browser
during the time the Snackbar was open, Autofill Assistant would revive
in the browser.

This happens, because the UI gets destroyed and recreated, losing all
knowledge about the Snackbar.

Now, if a Snackbar is showing when the switch happens, we execute the
callback immediately, which might shut down Autofill Assistant.

Bug: b/159784978
Change-Id: I600bc0de4796c9f65533d1627030963a83b02865
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302196
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#789447}