Mac: constrainFrameRect:toScreen: rework
When the screen passed to `constrainFrameRect:toScreen:` is nil,
AppKit does some gymnastics to select a screen, only defaulting
to the main screen if there's no result. We don't have access
to these calculations in an override, so we can't use the
screen's frame for heuristics when deciding whether to allow
the constrain to happen or not.

This change removes the previous logic in favor of trying
to test specifically for the large vertical jumps that cause
glitches when detaching tabs.

Bug: 1102925
Change-Id: Ic5653d5fd867da155605cd986ec80c5bb4afd755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298056
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789503}