Use code review in DOM distiller dist repo
Before this CL, this script builds DOM distiller, pushes the compiled
artifacts to the dom-distiller/dist repo, and generates a local commit
for rolling. The local commit needs to be uploaded as a CL.

In order to avoid pushing without reviewing, we want to generate a CL
in the dom-distiller/dist repo instead of directly pushing to it.

When reviewing the CL in the dom-distiller/dist repo, the reviewer
needs to run update_domdistiller_js.sh in validation mode to make sure
the artifacts in the CL are identical to local output.

Bug: None
Change-Id: Ia29c1686106c3516f341e4d826aca69842ac5668
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2304338
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789575}