Report update manifest download time using UMA statistics
We want to improve the installation time for force installed
extensions. For this, we need to first gather the statistics regarding
which installation stages are time consuming. We collect the UMA
statistics for the time taken to download and parse the update
manifest successfully.

Following histogram is added:
Extensions.ForceInstalledTime.DownloadingStartTo.
ManifestDownloadComplete

Bug: 1102369
Change-Id: I40137e7551216ebce5bb97046b1e1735184220b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2282588
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789613}