GMC: Always display playback controls LTR
Regardless of how the UI is laid out, media playback controls should be
laid out in a left-to-right manner, as suggested by the material.io
bidirectionality guide.

Additionally, the images for playback control buttons will no longer be
mirrored about their horizontal axes in an RTL layout. This includes
the play button and the PiP button. The PiP button is not mirrored
because the PiP display is always in the bottom right corner regardless
of UI layout.

Finally, after having worked with l10n and talking to UX engineers in
Israel, we have determined that the appropriate focus order for this
dialog should traverse the playback controls left-to-right first and
then focus the picture-in-picture button. This CL implements this order.

Bug: 1040021
Change-Id: I0a57ec7d879ab312ec1ce7bb592d8f79edceaa69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253134
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789662}