[omnibox] [rich-autocompletion] Unelide omnibox before setting focus.
With the 1-line UI, rich autocompletion adds a label adjacent to the
omnibox and then resizes the two as the autocompletion length changes.

On location bar focus, the omnibox is unelided; e.g. google.com is
unelided to https://www.google.com.

Before this CL, unelision occurs *after* the location bar is given focus
and its layout computed.

With this CL, unelision occurs *before* the location bar is given focus
so that it can reserve sufficient width for the unelided URL when
updating its layout.

Bug: 1062446
Change-Id: Id78d635255de69d9af97a5589a9e001eee3d0feb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2304666
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789698}