[Launcher impressions] Record ignores and don't record empty results
There's an extra kind of result action (impression, launch, abandon)
that we don't currently log: ignores. These happen when results are
shown (impression logged) but not used, in such a way that it's not
reasonable to log an abandon. We don't have much interest in ignores
on their own, but they're useful for validating that the metrics are
correct. Let's log them.

Bundled with the change: we're erroneously triggering impressions and
abandons sometimes when there are zero results of a kind displayed. This
adds an extra check.

Bug: 1097599
Change-Id: Ib02ab5a0adb8e986c29375cd022fabf24d31fc06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2301530
Reviewed-by: Rachel Wong <wrong@chromium.org>
Commit-Queue: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789805}