Fix Webapp.CheckServiceWorker.Time metric (CheckOfflineCapability)
Webapp.CheckServiceWorker.Time is used for the Finch experiment
(CheckOfflineCapability), but it always records a minimum value (0.01)
because the order of a subtraction was wrong. The result of a
subtraction became a minus value and it was rounded to a minimum value.

This CL fixes the order in a subtraction.

Bug: 965802, 1106693
Change-Id: Ia0e6172a34ba1e30e4c791e5e38a230aa312eaf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303933
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789822}