[Android] Remove redundant argument in AccountManagerTestRule.addAccount
This CL removes the redundant argument accountName in the method
AccountManagerTestRule#addAccount as this method is only responsible of
adding an account and its corresponding profile data to the fake
AccountManagerFacade and fake ProfileDataSource, so the accountName
always equals the one in ProfileData. We can make this easier to have
only one argument for this method.


Bug: 1094298
Change-Id: I7a6136bf1f0d2e41376828dd95ddd91921bbe013
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303398
Commit-Queue: Alice Wang <aliceywang@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789881}