[Autofill Assistant] Enable LocationBar test
It appears that waiting for the URL is flaky. This check is not really
required for the test, as we only want to see the error appearing once
the URL is committed (we can assume that the URL being entered into the
location bar will eventually load).

Bug: 1105146
Change-Id: I68ac2f5c5591dbcf8f85665429720152df9889b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2306090
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#789894}