[Autofill Assistant] Show error when switching back to a stopped AA tab
Currently when we have an error we go into the stopped state and
we destroy the controller. However this means that if we have a
tab showing an error message and we switch away from it and switch
back the error will not be showed again because we rely on the
controller to rebuild the UI on tab switched.

With this CL, when an error happens, we keep the controller and we
destroy it only when the UI is destroyed (by navigating away or tapping
on the close button).

Bug: b/152949732
Change-Id: I92d443ca433052a8a4e0fe7cb28e27463d752c6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2296080
Commit-Queue: Luca Hunkeler <hluca@google.com>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789918}