[headless] Disable headless protocol browsertests.
Various of the tests in the suite have been individually disabled due
to flakiness, but according to flake-portal most if not all of the tests
in the suite are flaky, so disable them all.

The individually disabled tests should be reviewed when re-enable this
suite, since at least some look to be instances of the same general
issue with tests flakily timing-out.

TBR=pfeldman

Bug: 1107396
Change-Id: Ie6540e92999503c2c83bd132f447860b4c2c8b45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2307216
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Wez <wez@chromium.org>
Auto-Submit: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789934}