Fix: Notify sync-prefs-change listeners for signed-in not syncing users
Before this CL, no SyncPrefs[1] were sent via the sync-prefs-changed
WebUI event if the user was signed-in but had Sync disabled. This CL
changes this, relaxing the early exit checks in PushSyncPrefs().
The change is safe because checking IsEngineInitialized() is enough to
guarantee the existence of the SyncUserSettings used to fill the JS
object.

The CL also fixes an aspect overlooked in crrev.com/c/1718425.
PushSyncPrefs() had the undocumented side-effect of setting the
configuring_sync_ member, which makes no sense when the method is called
in HandleSyncPrefsDispatch(). So the assignment is now moved out of
PushSyncPrefs() and copied to the other callsites.

[1] https://source.chromium.org/chromium/chromium/src/+/9cff09c426fb0436ce35f7c571f555e576534aaf:chrome/browser/resources/settings/people_page/sync_browser_proxy.js;l=94

Bug: 1106350
Change-Id: I9d3167d534957f10dcd878eda918bdd1510c2973
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302852
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#789942}