[WebOTP] Refactor sms service tests to not require a particular backend
Currently the sms_service_unittests only exercises the verification
backend which is not actually active by default since we are
currently using User Consent backend.

More than half the tests (10 out of 10) seems to apply to both backends.
This patch refactors the logic so that the backend specific part of
sms service are well encapsulated.

It then updates the test so these common tests are exercised with
both backends.

Changes:
 - Introduce UserConsentHandler with two specific implementation
    1. Noop: which does not show any UI to user and is used with user
       consent backend.
    2. PromptBased: which prompts user to give consent and is used with
       verification backend. This encapsulates previous OpenInfoBar,
       OnCancel, OnConfirm methods from sms service.
 - The consent handler is created on construction of the service and is
   used for its lifetime. This makes it possible to easily inject mock
   consent handlers for testing.
 - Update mock Service to use a simple Noop consent handler for common
   tests which don't involve consent.
 - Introduce a more specialized ServiceWithPrompt that uses a prompt
   based consent handler for tests specific to prompt based logic.
 - Add unit tests for consent handlers.
 - Minor renaming, comments to clarify the code further.


There is not expected to be any functional changes in this CL.


Change-Id: I411bf9d12cb7a2051e545ce6dd8a50fe37204326
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254447
Reviewed-by: Sam Goto <goto@chromium.org>
Commit-Queue: Majid Valipour <majidvp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789970}