[files-progress] Add new strings for transfer details.
There are 2 messages for time display as it may show time in 2 parts
like '1 hour 2 min' or 1 part like '2 hours' or '3 min'.

We already have Dismiss label but its ID and usage is for drive banner.
Added this new one as a general one to prevent confusion when
referencing in the code or changing the other one.

Bug: 953308
Change-Id: I7b7a2603e81ab8aea36a518d334546e1c4b8c905
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2306112
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Omid Tourzan <oto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789982}