[Paint Preview] Decouple *Controller classes from Mediator
Break cyclic dependency of PlayerFrameMediator with
- PlayerFrameScaleController
- PlayerFrameScrollController

BitmapState is tightly coupled to PlayerFrameMediator with good reason
so a cyclic dependency there is somewhat unavoidable.

This is achieved by having PlayerFrameCoordinator instantiate
everything and then introducing a new class that handles dispatch of
gestures. A helper for scaling is also added to break the two-way
interdependence that existed in the mediator.

Bug: 1099722
Change-Id: I13b258d973a5ff81cdf75505f59aabb02d66a6c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2295751
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Reviewed-by: Fred Mello <fredmello@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789996}