AW docs: mention CTS requires en-US locale
This amends documentation to mention CTS generally requires the en-US
locale. This is actually a requirement for
WebSettingsTest#testAccessAllowFileAccess(), and it's probably worth
calling out explicitly since some of the WebView team are based in the
United Kingdom (and en-GB is *not* compatible).

For more context, see
https://android-review.googlesource.com/c/platform/cts/+/1363422

Test: Upload to gerrit > open file > click "gitiles"
Change-Id: I4d8f197d9cf2ed259a3757650d042725448f2a94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2304555
Commit-Queue: Richard Coles <torne@chromium.org>
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790047}