Clipboard API: Move Feature Policy TestExpectation failures.
Sheriffs correctly disabled these Clipboard API Feature Policy tests,
that were failing on same-origin iframes not getting focus quickly
enough on MacOS. That said, while this issue wasn't completely expected,
we did know of an underlying issue that could lead to this, so move
these 2 test failures to be grouped with related cross-origin failures.

These test failures are due to focus() on Chrome not being synchronous,
as specified. These tests are not being immediately fixed, as the
synchronous focus() not existing on Chrome is not specified, so we
shouldn't add waiting for this async focus behavior until either:
 (1) this is specified (hopefully this will happen), or
 (2) focus() is sync on Chrome (unlikely given cross-origin security)

Bug: 1104125, 1105278
Change-Id: I91cc3c20eb7af4357669548e906bf0b7fc7f0298
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303286
Auto-Submit: Darwin Huang <huangdarwin@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790053}