[Switch Access] Wait to set listener on keyboard
Previously, Switch Access would try to set a listener on the keyboard
node, regardless of whether the node was found in the tree.

This change checks if the keyboard is found, and if not sets a callback
for when the node is found.

It also re-orders the calls in enterKeyboard() to reduce the times when
focus is drawn back to the text field after it has shifted to the
keyboard.

AX-Relnotes: n/a.
Bug: None.
Change-Id: I79663f1bc3297ce8e468e38f0364bd7e3a90f01a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302748
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790071}