Do not unblocklist a remotely disabled extension.
When the blocklist updates, the blacklist state
map may be empty or does not include the remotely
disabled extension. This means that not even a
"NOT_MALWARE" is sent for the extension. The
extension then gets put into no_longer_blocked
which is not what is expected.

Bug: 1107040
Change-Id: Ia58f0d7b831b54b02843d5c03f9179e84ec11ac5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2305598
Commit-Queue: Bettina Dea <bdea@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790076}