Use wheelTick in events/mouse-event-buttons-attribute.html
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. This CL we should replace
eventSender.mouseScrollBy with wheelTick which uses
gpuBenchmarking.smoothScrollByXY in
events/mouse-event-buttons-attribute.html.

Bug: 1047176
Change-Id: Ife50b732a6ec0733e84b034c07565a4a5138b51a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2173614
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790142}