cros: Disable toggle fullscreen when overview is starting up.
When a user has a windowed app that can be fullscreened and they toggle
overview, if they also toggle fullscreen fast enough it will cause
visual glitches with overview mode(no blur, no opaque background, etc.)

This CL addresses this by adding a check to HandleToggleFullscreen() to
see whether the overview is in its start animation. If it is, the toggle
fullscreen call is ignored. Otherwise, it proceeds normally.

Test: AcceleratorControllerTest.DontToggleFullscreenWhenOverviewStarts
Bug: 1094739
Change-Id: Ifff8afa39f238e2725220a0bbb4a92836ea1393d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2303380
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Jeremy Chinsen <chinsenj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790149}