Hacky fix for DoH SECURE mode timeouts
When configuring NetworkService to do DNS in SECURE mode, override the
configuration to allow up to two retries.  The retries themselves aren't
super important, but as the DNS stack lets the original request
continue in parallel with retries, this has the overall effect of
significantly increasing the effective timeout.

Bug: 1105138
Change-Id: Id8aec72a555c3a0e1bafdf16333d4efdc59b39d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308037
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Eric Orth <ericorth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790150}