[Shelf] Fix the flaky home button test case
My previous CL (https://crrev.com/c/2290766) fails to consider about the
situation where hotseat is disabled. It assumes that hotseat is always
enabled. Although hotseat is stable now, the flag to enable/disable
hotseat still exists. So this CL is created to protect the feature of
the old shelf.

This CL ensures that the clip rectangle is applied only when the
hotseat is enabled.

Bug: 1106253
Change-Id: I4f5b56fb6f155450bde562a0e9d20ed4649d9e44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2308032
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790212}