Cryptids: Injecting view into NTP logo space
This adds a GIF view that will eventually be used for cryptid rendering.

The general approach is:
* The existing logo is wrapped in a FrameLayout with a new ViewStub.
* When we get a callback indicating that no doodle is present, the NTP
  requests a drawable from the PCR and injects it into an ImageView
  inflated from the ViewStub.

Out-of-scope for this CL is to actually populate the GIF image data; I
have been testing locally with a res/raw file read into a byte array.

Change-Id: Ib5deae5322856c9d8e8a68f0fd46cdb2aaadbe79
Bug: 1061949
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2245888
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: sebsg <sebsg@chromium.org>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790226}