Report substages of the CRX installation
We need to improve on installation time for the force installed
extensions. For this, we first need to investigate which stage takes
more time. After CRX has been downloaded, it's installation happens
in several stages (verifying, unpacking etc), and now
InstallStageTracker tracks these stages so that we could get a better
idea that which stage is more time consuming. This is a
prework for adding UMA statistics for reporting time taken for each
stage.

Bug: 1102806
Change-Id: I1a484363ef06f8f04787ddb8086dbb401523206f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283607
Commit-Queue: Swapnil Gupta <swapnilgupta@google.com>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Oleg Davydov <burunduk@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790331}