Adapt tests for kDecoupleSyncFromAndroidMasterSync
This CL accomplishes two things.

1) Make sure test coverage is not lost when the feature is enabled.

crrev.com/c/2299353 hardcoded that many tests should consider
the feature as enabled/disabled. This CL makes sure that:
a) Tests that are agnostic to the feature don't specify this.
b) Most tests that do depend on the feature now use conditional setup/
expectations when appropriate.
c) Only tests that are specific to one state of the feature (enabled/
disabled) specify this.

2) Make sure all existing tests pass when the feature is enabled.

Bug: 1105795
Change-Id: I5825afe82041f7d91d7dcbf3d63f64ad74590ef7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2309699
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#790391}